Structure improvements: CSS files moved to /styles folder; added a /service for chatGPT API calls;
