+++ src/index.css
@@ -5,14 +5,6 @@
   --background-color: #131927;
   --primary-color: #001492;
   --secondary-color: #7532fb;
-  /* --tertiary-color: #000000; */
-  /* --quarternary-color: #000000; */
-  /* --btn-primary-color: #001492;
-  --btn-secondary-color: #7532fb;
-  --prompt-primary-color: #f42f5f;
-  --prompt-secondary-color: #8e25bf;
-  --response-primary-color: #fe8d5c;
-  --response-secondary-color: #fe4d57; */
   --btn-primary-color: #001492;
   --btn-secondary-color: #7532fb;
   --prompt-primary-color: #279cce;
